Fixes three small bugs

+++ code/game/objects/items/weapons/material/material_weapons.dm
@@ -5,7 +5,7 @@
 	health = 100
 	hitsound = 'sound/weapons/bladeslice.ogg'
 	gender = NEUTER
-	
+
 	throw_range = 7
 	w_class = ITEM_SIZE_NORMAL
 	sharp = 0
@@ -75,11 +75,13 @@
 
 /obj/item/weapon/material/proc/shatter(var/consumed)
 	var/turf/T = get_turf(src)
-	T.visible_message("<span class='danger'>\The [src] [material.destruction_desc]!</span>")
-	playsound(src, "shatter", 70, 1)
-	if(!consumed && drops_debris)
-		material.place_shard(T)
-	qdel(src)
+	if (T)
+		T.visible_message("<span class='danger'>\The [src] [material.destruction_desc]!</span>")
+		playsound(src, "shatter", 70, 1)
+		if(!consumed && drops_debris)
+			material.place_shard(T)
+	if (!QDELETED(src))
+		qdel(src)
 /*
 Commenting this out pending rebalancing of radiation based on small objects.
 /obj/item/weapon/material/process()
